Revert "Added to Git-Ignore"
