--- .gitignore
@@ -1,6 +1,0 @@
-MID_HUD.mel
-Pipeline_install.mel
-erase_virus.mel
-Update erase_virus_userScripts.py
-frases
-version.json
