+++ version.json
@@ -1,5 +1,5 @@
 {
-    "cams_free": {
+    "cams": {
         "version": "0.0.4",
         "changelog": [
             "=== 0.0.4 ===",
